Post handler that forwards submitted data as JSON
- renamed FormPostService in FormHandler
- changed api of form handler by only passing formData and endpointUrl
- fix issue with requestDispatcher.forward
