+++ bundles/core/src/main/java/com/adobe/cq/wcm/core/components/internal/models/v1/form/ContainerImpl.java
@@ -45,7 +45,6 @@ import com.adobe.cq.export.json.ExporterConstants;
 import com.adobe.cq.export.json.SlingModelFilter;
 import com.adobe.cq.wcm.core.components.internal.Utils;
 import com.adobe.cq.wcm.core.components.internal.form.FormConstants;
-import com.adobe.cq.wcm.core.components.internal.models.v1.AbstractComponentImpl;
 import com.adobe.cq.wcm.core.components.models.form.Container;
 import com.day.cq.wcm.api.Page;
 import com.day.cq.wcm.foundation.forms.FormStructureHelper;
@@ -59,7 +58,7 @@ import static com.day.cq.wcm.foundation.forms.FormsConstants.SCRIPT_FORM_SERVER_
        adapters = {Container.class, ContainerExporter.class, ComponentExporter.class},
        resourceType = {FormConstants.RT_CORE_FORM_CONTAINER_V1, FormConstants.RT_CORE_FORM_CONTAINER_V2})
 @Exporter(name = ExporterConstants.SLING_MODEL_EXPORTER_NAME, extensions = ExporterConstants.SLING_MODEL_EXTENSION)
-public class ContainerImpl extends AbstractComponentImpl implements Container {
+public class ContainerImpl implements Container {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ContainerImpl.class);
     private static final String PROP_METHOD_DEFAULT = "POST";
@@ -121,8 +120,11 @@ public class ContainerImpl extends AbstractComponentImpl implements Container {
     private void initModel() {
         FormStructureHelper formStructureHelper = formStructureHelperFactory.getFormStructureHelper(resource);
         request.setAttribute(FormsHelper.REQ_ATTR_FORM_STRUCTURE_HELPER, formStructureHelper);
-        request.setAttribute(FormsHelper.REQ_ATTR_FORMID, getId());
         this.action = Utils.getURL(request, currentPage);
+        if (StringUtils.isBlank(id)) {
+            id = FormsHelper.getFormId(request);
+        }
+        request.setAttribute(FormsHelper.REQ_ATTR_FORMID, getId());
         this.name = id;
         this.dropAreaResourceType = "wcm/foundation/components/responsivegrid/new";
         if (redirect != null) {
@@ -169,6 +171,11 @@ public class ContainerImpl extends AbstractComponentImpl implements Container {
     }
 
     @Override
+    public String getId() {
+        return this.id;
+    }
+
+    @Override
     public String getName() {
         return this.name;
     }
@@ -194,7 +201,7 @@ public class ContainerImpl extends AbstractComponentImpl implements Container {
         if (errorMessages != null && errorMessages.length > 0) {
             return Arrays.copyOf(errorMessages,errorMessages.length);
         }
-        return null;
+        return new String[]{};
     }
 
     @NotNull
