+++ bundles/core/src/test/java/com/adobe/cq/wcm/core/components/internal/models/v1/form/FormsHelperStubber.java
@@ -44,14 +44,18 @@ public class FormsHelperStubber {
             // set the body of all methods in the class to empty,
             // to remove any dependencies on impl classes.
             for (CtMethod ctMethod : ctClass.getDeclaredMethods()) {
-                ctMethod.setBody(null);
+                if (!ctMethod.getName().equals("getContentRequestParameterNames") &&
+                        !ctMethod.getName().equals("getFormId")) {
+                    ctMethod.setBody(null);
+                }
             }
             // remove the error causing static field declaration
             ctClass.removeField(ctClass.getDeclaredField(ERROR_FIELD));
             // remove the static initializer block calling new on impl class.
             ctClass.removeConstructor(ctClass.getClassInitializer());
             // defer getValues(..) to another method call so that we can manipulate/mock return values
-            ctClass.getDeclaredMethod("getValues").setBody("return com.adobe.cq.wcm.core.components.internal.models.v1.form.FormsHelperGetValuesStubMethod.get();");
+            ctClass.getDeclaredMethod("getValues")
+                    .setBody("return com.adobe.cq.wcm.core.components.internal.models.v1.form.FormsHelperGetValuesStubMethod.get();");
             // load the stubbed class
             ctClass.toClass();
         } catch (NotFoundException e) {
