+++ bundles/core/src/test/java/com/adobe/cq/wcm/core/components/internal/servlets/FormActionRpcServletTest.java
@@ -16,39 +16,50 @@
 package com.adobe.cq.wcm.core.components.internal.servlets;
 
 import java.io.IOException;
+import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 
 import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.http.osgi.services.HttpClientBuilderFactory;
+import org.apache.sling.api.request.RequestDispatcherOptions;
 import org.apache.sling.api.resource.ModifiableValueMap;
 import org.apache.sling.api.resource.Resource;
+import org.apache.sling.testing.mock.sling.servlet.MockRequestDispatcherFactory;
+import org.apache.sling.testing.mock.sling.servlet.MockSlingHttpServletRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.adobe.cq.wcm.core.components.context.CoreComponentTestContext;
-import com.adobe.cq.wcm.core.components.internal.form.FormPostServiceImpl;
+import com.adobe.cq.wcm.core.components.internal.form.FormHandlerImpl;
 import com.adobe.cq.wcm.core.components.internal.models.v1.form.FormsHelperStubber;
+import com.day.cq.wcm.foundation.forms.ValidationInfo;
 import com.github.tomakehurst.wiremock.WireMockServer;
+import com.google.common.collect.ImmutableMap;
 import io.wcm.testing.mock.aem.junit5.AemContext;
 import io.wcm.testing.mock.aem.junit5.AemContextExtension;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.*;
 import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 @ExtendWith({AemContextExtension.class, MockitoExtension.class})
-class FormRpcServletTest {
+class FormActionRpcServletTest {
 
-    private FormRpcServlet underTest;
+    private FormActionRpcServlet underTest;
 
     private WireMockServer wireMockServer;
     private int wireMockPort;
 
     public final AemContext context = CoreComponentTestContext.newAemContext();
-    private static final String TEST_BASE = "/form/form-post-service";
+    private static final String TEST_BASE = "/form/form-rpc-servlet";
     private static final String CONTENT_ROOT = "/content";
+
+    @Mock
+    private RequestDispatcher requestDispatcher;
 
     @BeforeEach
     void setUp() {
@@ -58,18 +69,46 @@ class FormRpcServletTest {
         wireMockPort = wireMockServer.port();
         setupStub();
         context.registerService(HttpClientBuilderFactory.class, HttpClientBuilder::create);
-        context.registerInjectActivateService(new FormPostServiceImpl());
-        underTest = context.registerInjectActivateService(new FormRpcServlet());
+        context.registerInjectActivateService(new FormHandlerImpl());
+        context.request().setRequestDispatcherFactory(new MockRequestDispatcherFactory() {
+            @Override
+            public RequestDispatcher getRequestDispatcher(String path, RequestDispatcherOptions options) {
+                return requestDispatcher;
+            }
+
+            @Override
+            public RequestDispatcher getRequestDispatcher(Resource resource, RequestDispatcherOptions options) {
+                return requestDispatcher;
+            }
+        });
+        underTest = context.registerInjectActivateService(new FormActionRpcServlet());
         FormsHelperStubber.createStub();
     }
 
     @Test
-    void testDoPost() throws ServletException, IOException {
+    void testDoPostWithSuccess() throws ServletException {
+        MockSlingHttpServletRequest request = context.request();
+        request.setParameterMap(ImmutableMap.of("text", "hello"));
+        request.setAttribute("cq.form.id", "new_form");
         Resource resource = context.currentResource("/content/container");
         ModifiableValueMap modifiableValueMap = resource.adaptTo(ModifiableValueMap.class);
         modifiableValueMap.put("formEndPointUrl", "http://localhost:" + wireMockPort + "/form/endpoint");
-        underTest.doPost(context.request(), context.response());
+        underTest.doPost(request, context.response());
         assertEquals(302 , context.response().getStatus());
+    }
+
+    @Test
+    void testDoPostWithError() throws ServletException {
+        MockSlingHttpServletRequest request = context.request();
+        request.setParameterMap(ImmutableMap.of("text", "hello"));
+        request.setAttribute("cq.form.id", "new_form");
+        Resource resource = context.currentResource("/content/container");
+        ModifiableValueMap modifiableValueMap = resource.adaptTo(ModifiableValueMap.class);
+        modifiableValueMap.put("formEndPointUrl", "http://localhost:" + wireMockPort + "/form/nonExistingEndpoint");
+        underTest.doPost(context.request(), context.response());
+        assertEquals(200 , context.response().getStatus());
+        ValidationInfo validationInfo = ValidationInfo.getValidationInfo(request);
+        assertNotNull(validationInfo);
     }
 
     private void setupStub() {
