+++ bundles/core/src/test/resources/form/container/exporter-container_350773202.json
@@ -2,6 +2,7 @@
     "method"                 : "POST",
     "enctype"                : "multipart/form-data",
     "action"                 : "/core/content/coretest/demo-page.html",
+    "errorMessages"          : [],
     "resourceTypeForDropArea": "wcm/foundation/components/responsivegrid/new",
     ":itemsOrder"            : [],
     ":type"                  : "core/wcm/components/form/container/v1/container",
